Production check: fix 30 confirmed findings (1 CRITICAL, 29 MEDIUM) across C++/Python/tests
CRITICAL (1):
- sgs_params.cpp: Fix default SGS kriging type KRIG_ORDINARY -> KRIG_SIMPLE (GSLIB requirement)

C++ Numerical Correctness (12):
- gauss_solver.cpp: Add residual norm check after back-substitution for solution quality
- my_kriging_weights.h: Compute kriging variance in 6 Cholesky-failure fallback paths (2 OK + 4 SK)
- sugarbox_indexed_neighbour_lookup.h: Add octant spatial-diversity enforcement (GSLIB noct equivalent)
- indicator_kriging.h: Fix indicator encoding mismatch (exclusive PMF -> cumulative CDF prefix sum)
- sequential_simulation.h: Use LVM prior mean in kriging-failure fallback instead of N(0,1)
- sequential_indicator_simulation.cpp: Add probability clamping for 2-category SIS path
- sequential_indicator_simulation.cpp: Add property/grid size validation to LVM SIS entry point

C++ Random Path & Seed (5):
- path_random_generator.cpp: Replace LCG with Fisher-Yates shuffle (uniform permutations)
- path_random_generator.h: Accept int64_t seed (was int, causing silent truncation)
- api.cpp/api.h/hpgl_core.h: Propagate int64_t seed through entire SIS chain (no truncation)
- sequential_simulation.h: Add negative seed validation
- random_node_enumerator.h/cpp: Accept int64_t seed in constructor

C++ API/IO Hardening (5):
- api.cpp: Throw on partial remap table mismatch (was silently discarding caller values)
- api.cpp: Add output shape volume validation to hpgl_lvm_kriging
- api.cpp: Add indicator_count==2 validation to hpgl_median_ik
- load_property_from_file.cpp: Remove isalpha() restriction on property names
- load_property_from_file.cpp: Throw immediately on first bad token in load_doubles_into_vector

Python API (4):
- hpgl_wrap.py: Change SIS ctypes seed from c_int to c_int64 (match C API)
- geo.py: Align undefined_value comparison to exact == (match C++ fast reader)
- geo.py: Add indicator_count==2 validation to median_ik()
- sis.py: Remove dead min_neighbours parameter from sis_simulation()

Documentation (1):
- CHANGELOG.md: Remove false claim about use_regions/region_size features

Tests (3 files, +10 new tests):
- test_kriging_complete.py: Add analytical OK interpolation tests, tighten bounds, add NaN/Inf checks
- test_simulation_complete.py: Tighten variance bounds 4x->1.5x, add mean/variogram validation
- test_cvariogram.py: Add analytical gamma(h) assertions and tol_distance boundary tests

Total: 22 files, +754 -396 lines. Build: 30/30 targets. Tests: 763 passed, 2 skipped, 0 failed.

+++ src/geo_bsd/hpgl/api.h
@@ -292,7 +292,7 @@ hpgl_sis_simulation(
 		hpgl_ind_masked_array_t * data,
 		hpgl_ik_params_t * params,
 		int indicator_count,
-		int seed,
+		int64_t seed,
 		hpgl_ubyte_array_t * simulation_mask);
 
 HPGL_API void
@@ -301,7 +301,7 @@ hpgl_sis_simulation_lvm(
 		hpgl_ik_params_t * params,
 		hpgl_float_array_t * mean_data,
 		int indicator_count,
-		int seed,
+		int64_t seed,
                 hpgl_ubyte_array_t * simulation_mask,
 		int use_correlograms);
