+++ src/geo_bsd/hpgl/path_random_generator.h
@@ -1,7 +1,9 @@
 #ifndef __PATH_RANDOM_GENERATOR_H__907328AB_FE6E_437B_9725_65FD1119284A__
 #define __PATH_RANDOM_GENERATOR_H__907328AB_FE6E_437B_9725_65FD1119284A__
 
+#include <cstdint>
 #include <memory>
+#include <vector>
 
 namespace hpgl
 {
@@ -11,8 +13,8 @@ namespace hpgl
 		std::shared_ptr<Impl> m_impl;
 	public:
 		path_random_generator_t();
-		void init(int size, int seed);
-		path_random_generator_t(int size, int seed);
+		void init(int size, int64_t seed);
+		path_random_generator_t(int size, int64_t seed);
 		~path_random_generator_t();
 
 		int next();
